requirements.txt
click==6.7
docutils==0.12
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.3.2
futures==3.0.4
gunicorn==19.4.1
itsdangerous==0.24
Jinja2==2.10
lockfile==0.12.2
MarkupSafe==1.0
mysql-connector-python==2.0.4
pika==0.10.0
pymongo==3.2
python-daemon==2.1.1
python-dateutil==2.5.0
requests==2.9.1
six==1.11.0
Werkzeug==0.14.1
wheel==0.24.0
